Dealt with Warnings on Compiling -- Part 1 49 Warnings  a. -Winconsistent-missing-destructor-override (2)  b. -Wextra-semi (3)  c. -Wsign-conversion (22)  d. -Wimplicit-int-float-conversion (10)  e. -Wfloat-conversion (8)  f. -Wimplicit-float-conversion (4)
Compiler: clang version 10.0.0-4ubuntu1
Target: x86_64-pc-linux-gnu
Note: I applied "unsigned int" rather than "uint32_t" as long as the conversion isn't for "lastPositionData.beatUnit" in "juce_LV2_Wrapper.cpp".

+++ modules/juce_audio_plugin_client/LV2/juce_LV2_Wrapper_Exporter.cpp
@@ -331,7 +331,7 @@ static const String makePluginFile (AudioProcessor* const filter, const int maxN
 
         text += "        a lv2:InputPort, lv2:ControlPort ;\n";
         text += "        lv2:index " + String (portIndex++) + " ;\n";
-        text += "        lv2:symbol \"" + nameToSymbol (paramName, i) + "\" ;\n";
+        text += "        lv2:symbol \"" + nameToSymbol (paramName, (unsigned int) i) + "\" ;\n";
 
         if (paramName.isNotEmpty())
             text += "        lv2:name \"" + paramName + "\" ;\n";
@@ -445,7 +445,7 @@ static const String makePresetsFile (AudioProcessor* const filter)
             else
                 preset += "    [\n";
 
-            preset += "        lv2:symbol \"" + nameToSymbol (parameters[j]->getName (1024), j) + "\" ;\n";
+            preset += "        lv2:symbol \"" + nameToSymbol (parameters[j]->getName (1024), (unsigned int) j) + "\" ;\n";
             preset += "        pset:value " + String::formatted ("%f", safeParamValue (parameters[j]->getValue())) + " ;\n";
 
             if (j+1 == filter->getParameters().size())
